# .static/swagger/swagger.yaml
definitions:
...
  api.createTeamRequest:
    properties:
      name:
        type: string
    type: object
  api.inviteRequest:
    properties:
      email:
        type: string
      role:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/teams:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Idempotency key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      - description: Create team
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create team
      tags:
      - teams
  /api/v1/teams/{id}/invite:
    post:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Idempotency key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      - description: Invite request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.inviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Invite user by email
      tags:
      - teams
swagger: "2.0"